+++ CHANGELOG.md
@@ -7,3 +7,5 @@
 01. :gift: Added class `dev.orne.test.rnd.AbstractGenerator`
 01. :gift: Added interface `dev.orne.test.rnd.TypedGenerator`
 01. :gift: Added class `dev.orne.test.rnd.AbstractTypedGenerator`
+01. :gift: Added exception `dev.orne.test.rnd.GeneratorNotFoundException`
+01. :gift: Added class `dev.orne.test.rnd.Generators`